Update GordonDialogBox
Update to Typescript for better prop typing, and fix some styles to be more consistent with the rest of the site.
