--- src/components/GordonDialogBox/GordonDialogBox.module.scss
@@ -1,18 +1,0 @@
-@import '../../vars';
-
-.gc360_gordondialogbox {
-  display: grid;
-  grid-template-rows: auto auto auto;
-  overflow-y: auto;
-  &_title {
-    background-color: $primary-blue;
-    color: $neutral-white;
-    text-align: center;
-  }
-  &_content :global(.MuiDialogTitle-root) {
-    padding: 8px 20px 8px 16px;
-  }
-  &_actions {
-    justify-content: space-evenly;
-  }
-}
